11-proyecto_whisper/requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.4.2
Eel==0.16.0
filelock==3.13.1
fsspec==2023.10.0
future==0.18.3
futures==2.2.0
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai-whisper @ git+https://github.com/openai/whisper.git@fcfeaf1b61994c071bba62da47d7846933576ac9
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pydub==0.25.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pymongo==4.6.0
pyparsing==3.1.1
pywin32-ctypes==0.2.2
regex==2023.10.3
requests==2.31.0
sympy==1.12
tiktoken==0.5.1
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
whichcraft==0.6.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
